notificator-app/requirements.txt
confluent-kafka==1.5.0
python-telegram-bot==13.11